wip: messing with build options for wasm cgo
